refactor(io): extract InputView and ResultView, slim down main()
